+++ CHANGES.md
@@ -2,6 +2,8 @@ Change Log
 ==========
 
 #### next release (8.1.10)
+
+* Fix `CswCatalogGroup` XML types
 * Added `MAINCODE` aliases for all ABS Statistical Area regions that were missing them.
 * Fixed `superGet` replacement in webpack builds with babel versions `7.16.0` and above.
 * [The next improvement]